Major top level redesign - first draft - inevitably broken
