+++ package.json
@@ -21,6 +21,7 @@
   "dependencies": {
     "firebase": "^9.6.8",
     "firebaseui": "^6.0.1",
+    "page": "^1.11.6",
     "sirv-cli": "^1.0.0",
     "tippy.js": "^6.3.7"
   }
