+++ src/pages/pagesstore.js
@@ -1,0 +1,3 @@
+import { writable } from 'svelte/store';
+
+export const currentPage = writable(null);
